+++ koala/trpc-routes/turbine/util.ts
@@ -1,0 +1,6 @@
+export function clean(input: string[]) {
+  return input
+    .map((item) => item.trim())
+    .filter((x) => x.length > 1)
+    .sort();
+}
